Use ints instead of floats to compute the penalties to have commutative results

+++ layout_generator/penalties/penalize_digraphs.hpp
@@ -4,12 +4,12 @@
 #include "layout_generator/penalties/ngraphs.hpp"
 
 namespace finger_tracking::penalties {
-inline float penalizeSameFingerTwice(KeyPress const& current, KeyPress const& previous) {
+inline int64_t penalizeSameFingerTwice(KeyPress const& current, KeyPress const& previous) {
     bool isSameFinger = current.fingerDesc() == previous.fingerDesc();
     return current.keyRef != previous.keyRef && isSameFinger ? 4.f : 0.f;
 }
 
-inline float penalizeLongJumpSameHand(KeyPress const& current, KeyPress const& previous) {
+inline int64_t penalizeLongJumpSameHand(KeyPress const& current, KeyPress const& previous) {
     bool isSameHand   = current.side() == previous.side();
     bool crossHomeRow = (current.row() > Row::Home && previous.row() < Row::Home)
                      || (current.row() < Row::Home && previous.row() > Row::Home);
@@ -17,7 +17,7 @@ inline float penalizeLongJumpSameHand(KeyPress const& current, KeyPress const& p
     return isSameHand && crossHomeRow ? 1.f : 0.f;
 }
 
-inline float penalizeLongJumpSameFinger(KeyPress const& current, KeyPress const& previous) {
+inline int64_t penalizeLongJumpSameFinger(KeyPress const& current, KeyPress const& previous) {
     bool isSameFinger = current.fingerDesc() == previous.fingerDesc();
     bool crossHomeRow = (current.row() > Row::Home && previous.row() < Row::Home)
                      || (current.row() < Row::Home && previous.row() > Row::Home);
@@ -25,7 +25,8 @@ inline float penalizeLongJumpSameFinger(KeyPress const& current, KeyPress const&
     return isSameFinger && crossHomeRow ? 1 : 0.f;
 }
 
-inline float penalizeLongJumpConsecutiveFingers(KeyPress const& current, KeyPress const& previous) {
+inline int64_t penalizeLongJumpConsecutiveFingers(KeyPress const& current,
+                                                  KeyPress const& previous) {
     bool areConsecutiveFinger
         = current.side() == previous.side() && std::abs(current.finger - previous.finger) == 1;
 
@@ -35,7 +36,7 @@ inline float penalizeLongJumpConsecutiveFingers(KeyPress const& current, KeyPres
     return areConsecutiveFinger && crossHomeRow ? 1.f : 0.f;
 }
 
-inline float penalizeFingerTwist(KeyPress const& current, KeyPress const& previous) {
+inline int64_t penalizeFingerTwist(KeyPress const& current, KeyPress const& previous) {
     bool isSameHand         = current.side() == previous.side();
     bool ringDownAfterPinky = current.finger == Finger::Ring && previous.finger == Finger::Pinky
                            && current.row() < previous.row();
@@ -47,7 +48,7 @@ inline float penalizeFingerTwist(KeyPress const& current, KeyPress const& previo
              : 0.f;
 }
 
-inline float penalizeRollOut(KeyPress const& current, KeyPress const& previous) {
+inline int64_t penalizeRollOut(KeyPress const& current, KeyPress const& previous) {
     bool isSameHand = current.side() == previous.side();
     bool isRollOut  = current.finger != Finger::Thumb && previous.finger != Finger::Thumb
                   && current.finger > previous.finger;
@@ -55,7 +56,7 @@ inline float penalizeRollOut(KeyPress const& current, KeyPress const& previous) 
     return isSameHand && isRollOut ? 1 : 0.f;
 }
 
-inline float penalizeRollIn(KeyPress const& current, KeyPress const& previous) {
+inline int64_t penalizeRollIn(KeyPress const& current, KeyPress const& previous) {
     bool isSameHand = current.side() == previous.side();
     bool isRollIn   = current.finger < previous.finger;
